Devices.Device: add pause frames properties and methods
The Device class now provides following properties:
rx_pause_frames
tx_pause_frames

The property value is True if the pause frame setting is on or False
if it is turned off. The same applies when setting the property value.

The configuration of the pause frames is automatically restored to the
state before running a recipe.

Version 2:

Ethtool returns non-zero error code when the pause frames have the same
values as the values to be configured. To handle this I added a return
code check for this specific case.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>