lnst.Slave.InterfaceManager: handle RTM_DELLINK with family PF_BRIDGE properly
If a bridge device (t_br0) is removed and another device (e.g. t_vlan0)
references it as IFLA_MASTER (ie. the device was previously added to the bridge),
an RTM_DELLINK is sent by kernel to userspace for t_vlan0 device before the
bridge device is removed.

The InterfaceManager handles all RTM_DELLINK as a notification of removal of
a device. In this case however, the RTM_DELLINK has a special meaning that the
link between t_vlan0 and t_br0 does not exist anymore, and is indicated by
PF_BRIDGE (value 7) family in the RTM_DELLINK message.

Without this patch, upon delivery of such message, the t_vlan0 device is
removed from the InterfaceManager's device list and marked as deleted. Later
on during the deconfiguration, the actual removal (ip link delete) of the
device is skipped because LNST thinks it has been already deleted.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>