BaselineFlowAverageEvaluator: implement metric selection
It may sometimes be useful to ignore some metrics during evaluation. At
the moment this is intended to be used with TRexFlowMeasurement where
generator and receiver cpu stats are identical (because both generator
and receiver are often on the same host). And the generator throughtput
results are often less stable and less important.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>