VirtOvsVxlanRecipe: fix common hw config
The VirtOvsVxlanRecipe was missed during common hw config refactoring
and the old "hw_config" method was forgotten.

The recipe does inherit the correct subconfig mixin class however it's
not called due to this old hw_config method.

This patch removes the method and adds the missing device list
properties used by the common hw subconfig mixins.

The adaptive coalescing one is skipped for now as I have not found it
being used in the legacy version of the recipe.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>