Recipes.ENRT.ConfigMixins.BaseHWConfigMixin: do not initialize config if dev_list is empty
If a specific config mixin's dev_list property is left unset, the mixin's description
would be silently discarded. In such cases the mixin should report that configuration
of the feature has been skipped. That should be similar to the case when the parameter
controlling the mixin is not specified.

Fixes issue 199

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>