RecipeResults.py: Stop excluding RemoteDevices from export
Previously we were excluding RemoteDevice from export due to a number of
its attributes containing references to low level PYCapsule and socket
stuff.

I took another look at this to be more precise about what is excluded.

Also, with Ondrej help I was able to find the cause of an issue I a ran
into previously where an infinite loop of calls to
RemoteDevice.__setattr__ would occur when trying to unpickle a exported
run.

Signed-off-by: Perry Gagne <pgagne@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>