Recipes.ENRT.ConfigMixins: add DisableIdleStatesMixin
This mixin class extends BaseEnrtRecipe with ability to control the CPU idle
states before running the tests.

The class adds 'minimal_idlestates_latency' parameter that is passed as the
latency argument (-D) of the 'cpupower idle-set' that will disable all idle
states with an equal or higher latency than the specified value.

If the value is 0 this will effectively disable all CPU idle states.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>