Recipes.ENRT.ConfigMixins: add DisableTurboboostMixin
This mixin class provides recipe parameter 'disable_turboboost' that can be
used to disable CPU turboboost before running the tests.

It uses intel_pstate sysfs interface to disable the CPU feature and so
it is usable only by systems with Intel CPUs.

To use it pass 'disable_turboost=True' parameter when creating an instance of
BaseEnrtRecipe.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>