Recipes.ENRT.BaseEnrtRecipe: add DisableTurboboost and DisableIdleStates mixins
All of the tests that inherit from BaseEnrtRecipe will make use of these mixins
so instead of adding them to individual recipes they are included in the base
class.

To control the behaviour on a specific recipe level the recipes can override
the disable_turboboost_host_list and disable_idlestates_host_list properties
and relevant mixin recipe parameters.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>