RecipeCommon.PacketAssert: check module return value
Before returning any sort of PacketAssert "recipe template" result we
need to check the actual PacketAssert test module success or failure, if
the remote job returns a failure this means that it failed to measure
anything meaningful and that there are no results available.

Because of this we should return an empty object that tells the
evaluate_and_report method to skip a normal evaluation and just
immediately report a failing result.

This fixes a traceback crashing issue where trying to read "p_recv" was
not possible when the PacketAssert test module failed.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>