Revert "add sleep between start of iperf server and client"
This reverts commit a961862ff4734b609d7b6325bc2555829833f906.

Some issues popped up when this patch was used.

First issue was with the CPUStatMeasurement that is used in ENRT recipes.
The 2 seconds delay caused the measurement to include samples when the CPUs
were idle and reported 1-2% lower utilization.

Second issue was a result of iperf bug [1], when the iperf measures CPU
utilization from the moment the iperf server program is started, again
this includes 2 seconds of idling.

Until both issues are dealt with, the simplest solution is to revert the
patch.

[1] https://github.com/esnet/iperf/issues/1076

Signed-off-by: Jan Tluka <jtluka@redhat.com>