RecipeCommon.Perf.Measurements.StatCPUMeasurement: sort hosts
Sort the hosts where we start the CPUStatMonitor jobs, this is because
the provided `self._conf` object is a set() and this means that
sometimes the jobs get created in different orders.

This is ok wrt. validity of the jobs, however it makes it more difficult
for humans to compare result logs when the measurement results are
different each execution.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>