Recipes.ENRT.BaseEnrtRecipe: remove perf measurement generation
The generation of combinations of perf measurements is being split off
from the BaseEnrtRecipe class to make it easier to override in derived
classes.

The functionality is now instead available in the
Recipes.ENRT.MeasurementGenerators package.

We keep the evaluator registration in this class for now as we haven't
found a better place for the evaluator properties yet. However we adjust
it so that it doesn't access the "*_tool" parameters anymore and instead
adds evaluators from "*_evaluators" properties based on the base types
of measurements.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>