refactor lnst.Recipes.ENRT.BaseEnrtRecipe
Some of the subconfig and perf mixins are not relevant on the generic
BaseEnrtRecipe level and should be separated into a more specific
derived class. This is relevant for example for ovs+dpdk based recipes
where we don't want to do any device specific configuration (as devices
are reconfigured to use a different driver module than standard).

For now this includes the move of some of the base classes into the
BaremetalEnrtRecipe and VirtualEnrtRecipe that are then reused in the
specific recipes.

Eventually similar steps should be done with the ovs+dpdk type recipes.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>