lnst.RecipeCommon.Perf: Evaluators refactor to add parent recipe configuration
This commit adds "recipe_conf" parameters to the Perf.Evaluators code.

The value of this parameter is passed to the Evaluators code from the
Perf.Recipe::perf_evaluate method which provides the reference to a
recipe configuration that was stored in the *results* that are being
evaluated.

These results should be a *snapshot* of the recipe configuration that
was current at the point in time when the measurement was being run.

As such the "recipe_conf" in the evaluator code can be used to inspect
the *historical* state of the recipe configuration. This is important to
properly understand the result for the purposes of the evaluation.

It is also important to understand that this "recipe_conf" object is of
type lnst.RecipeCommon.Perf.Recipe.RecipeConf so this refers directly to
the configuration of the *Perf.Recipe*, not the overall configuration of
the entire recipe as it is.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>