RecipeCommon.Perf.RecipeConf: add parent recipe reference
This property can be used to reference the "full" recipe configuration
from the Perf.RecipeConfiguration - the main use case here is that
during evaluation we want to be able to inspect the *snapshot* of the
full configuration of the recipe at the time when the measurement was
taken (even if the evaluation is done later).

Because of this the initialization of the property from the
BaseEnrtRecipe is done via a "copy.deepcopy" call to ensure that the
full snapshot is stored.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>