lnst.RecipeCommon.Perf.Recipe: report and evaluate aligned results
Instead of using the whole set of samples from the measurements where
some of the samples may be irrelevant since each measurement can start
at different time, an aligned set of samples is generated to report and
evaluate.

The alignment is done on all measurements done within an iteration.
A maximum timestamp of the first samples and the minimum timestamp of
the last samples of each measurement is retrieved from the measurements.
A copy of RecipeResults data with any samples not aligned within the
timestamp range removed is generated and used for reporting and evaluation.

v2:
- use recipe_conf.iterations for retrieval of iterations count
- made list comprehension when collecting measurement results per iteration
  more readable

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>