Recipes.ENRT: move {idlestates,turboboost}_host_list properties
Because the mixin classes for idlestates and turboboost disabling were
moved to be inherited into the Baremetal and Virtual enrtrecipes base
classes, the host list definition in BaseEnrtRecipe was being overriden
by the higher placed base definition in the original mixin classes like
this:

    SimpleNetworkRecipe -> Mixin -> BaseEnrtRecipe

The base definition of the "host_list" just returns [], and doesn't call
super so the definition from BaseEnrtRecipe was never called. This
effectively disabled the mixins in all the enrt recipes that previously
used them.

Moving these to the new base classes of BaremetalEnrtRecipe and
VirtualEnrtRecipe fixes the issue.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>