*Measurement: add proper timestamp init where missing
We shouldn't use "None" for initialization of timestamps as it breaks
any time alignment/time slice related code that works with these
results.

Use current timestamp of the controller (code that creates the
PerfResult object) as "good enough" in cases where no timestamp is
returned from the Test tool used.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>