Perf.Results: refactor timestamping and implement time_slice()
Extending the PerfResult object api with the following
methods/properties:

* start_timestamp
* end_timestamp
* time_slice(start, end)

These unify the api for PerfInterval, SequentialPerfResult and
ParallelPerfResult.

This also means that "timestamp" itself in PerfInterval is renamed to
"start_timestamp" to avoid duplication.

The "time_slice" method returns an object of the same type
(PerfInterval, SequentialPerfResult, ParallelPerfResult or other class
derived from PerfResult) that is "sliced" to only contain measurement
data from the selected timestamp period restricted by the "start" and
"end" parameters.

The implementation depends on the specific class:
* PerfInterval serves as the smalles unit and creates a new PerfInterval
  object that has a shorter *duration* (intersection of current
  start-end and requested start-end), and has the *value*, adjusted by
  the ration of new_duration/old_duration. This ensures that the
  "average" calculated from this PerfInterval is consistent after the
  time_slice
* PerfList derivatives (SequentialPerfResult and ParallelPerfResult),
  recursively call "time_slice" in all the individual items of the
  sequence.

In case an empty slice would be created (mismatch of start-end
intervals), an EmptySlice exception is thrown.

This commit also adjusts the PerfRecipe "align_data" related code to use
this new refactored code.

v2:
* fixed Perf.Recipe time_aligned_results method - in case alignment
  isn't possible (broken measurement, or mismatched start-end time
  intervals) we should log errors and fall back the evaluation to the
  unaligned data

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>