IperfFlowMeasurement: fix parse_job_cpu duration get, again
One more fix for accessing the test duration when parsing test results
for cpu utilization of the iperf process.

The json dictionaries are inconsistent between tcp/udp stream on how
they report the end of test data - udp reports only a "sum" dictionary,
tcp only reports a "sum_sent" and "sum_received" dictionaries. So
instead we look at the test start and the *requested* test duration. Not
as precise as the actual duration of the test but probably good enough
considering how the duration is used in this case...

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>