Recipes.ENRT.SimpleMacsecRecipe: fix flow combination generation
The SimpleMacsecRecipe was overriding the generate_flow_combinations
method originally defined in BaseEnrtRecipe because there was no
separation of flow_endpoints yet.

This method override also contained a bug up until now - the
"ips_filter" was only using "family" filtering, but for ipv6 it should
also have used "is_link_local=False". This lead to undeteremenistic
behaviour when simetimes the link local address was used for the flow
measurement.

The solution is to remove this method and us the inherited method from
the IperfMeasurementGenerator class that has since been improved to use
"generate_perf_endpoints" which is added to the recipe instead.

I also added a wait_tentative_ips() call after calling "up()" on both
the eth and the macsec device in "apply_sub_configuration()" as this can
help avoid race condition issues when an ipv6 address isn't ready.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>