Recipes.ENRT.SimpleMacsecRecipe: macsec_settings renamed and made into a parameter, remove None value
The generate_sub_configurations method iterates over this list and
generates two subconfigurations leading to two perf measurements in a
single recipe run. Making this into a recipe parameter means we're able
to differentiate these two perf measurements in external databases.

I also removed the "None" value and the associated conditions. This was
originally used in the lnst-legacy recipe implementation to check the
functionality of the underlying connection but it doesn't really make
sense as a part of the recipe.

Finally, the parameter was renamed to "macsec_encryption" to make it
more self explanatory.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>