ShortLivedConnectionsRecipe.py: Update to use Neper
When ShortLivedConnectionsRecipe was originally ported to LNST next it was trying to use iPerf
However, iperf does not support Request/Response (RR) style tests.

The lnst-legacy version used netperf, which did support RR tests, but we wanted to use a different tool
that was easier to wrap.

This patch creates a new super class LatencyEnrtRecipe that uses the
Neper (https://github.com/google/neper) support from earlier patches.

ShortLivedConnectionsRecipe is a implementation of this class.

Signed-off-by: Perry Gagne <pgagne@redhat.com>

Edit from olichtne:
* removed the commented out perf_parallel_streams parameter from the
  ShortLivedConnectionsRecipe as its only purpose would be to override
  the inherited default value. As the parameter support is still missing
  and we've never actually run this test in the multiflow configuration
  (even in legacy xml recipes), this default value makes no sense.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>