Recipes.ENRT.ConfigMixins.DevInterruptHWConfigMixin: change dev_intr_cpu to ListParam
This is a change required by parallel iperf test. To achieve reproducible
measurements it's required to turn off irqbalance and spread the NIC IRQs
evenly on the available cpus.

The current recipe parameter dev_intr_cpu can specify only one cpu.

This patch changes the type of dev_intr_cpu parameter to ListParam.
The value is a list of cpus that will be used for pinning the NIC IRQs.

For the original behaviour a user simply needs to change
	dev_intr_cpu=0
to
	dev_intr_cpu=[0]
and the recipe will run as before.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>