Tests.PacketAssert: fix stderr handling
Tcpdump always prints information to stderr, the "ignore_exprs" regex
detection of "real" errors vs debug information is insufficient due to
this potentially changing at any point from tcpdump. An example of that
is that on RHEL8 we now also see the following line:

        dropped privs to tcpdump

Adjusted the PacketAssert module to only report PASS/FAIL based on the
return code of tcpdump (if not 0 then there's some issue). Text in
stderr should also always be debug logged, just in case it could be
useful later.

I also adjusted how the process is "finished" after an interrupt is
received so that the returncode is filled in properly (after the
communicate call).

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>