Tests.PacketAssert: continuosly read tcpdump stdout
When PacketAssert is used for longer or larger streams the buffer for
the pipe file used for stdout can be fully filled and tcpdump will get
stuck on the write syscall. After that when the SIGINT comes from the
recipe this syscall is interrupted and tcpdump quits with:

    tcpdump: Unable to write output: Interrupted system call

The fix in this patch is in two parts:
* use the '-U' argument which makes the tcpdump output packet buffered -
  fflush is called after each packet is analyzeed and printed.
* rewrite the wait_for_interrupt use with custom SIGINT handling that
  continuosly reads lines from stdout or stderr so that the buffer is
  never full

Eventually similar functionality may be useful for other Test modules as
well, at that point this should be refactored.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>