NeperFlowMeasurement: fix cpu usage PerfInterval calculation
The original calculation creates the ratio of the "used" cpu time vs the
elapsed real time, however to create a percentage we need to multiply by
100. Without this fix the reported values were always between 0 and 1
would seem confusing in the context of other cpu metrics where a 0 to
100 value is reported.

At the same time, the PerfInterval value should be multiplied by the
actual duration of the interval so that the average calculation makes
sense. This is required for metrics that don't represent a useful
"cumulative" unit/metric - throughput is a metric that accumulates
transferred bytes over time, however percentage cpu utilization doesn't
accumulate over time.

Without this fix, sometimes a short start or end interval (e.g. duration
of 0.00005 seconds) would end up confusing and poisoning the calculation
of the entire Sequence of Intervals.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>