Device: add a wait metchanism to _ip_add_one method
Unfortunatelly when an ip address is configured via netlink (a create ip
address message is sent) we don't receive a response message that would
indicate success or failure.

Because of that we've sometimes seen race condition issues with
configuring ipv6 addresses (that take slightly longer to create) on the
controller that was trying wait for the "tentative" state to disappear
from the ip addresses (because the address wasn't created yet there was
nothing to wait for).

By adding a wait loop here the race condition disappears - either the ip
address appears or an exception is raised indicationg a potential
problem with the configuration.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>