Controller.Machine: fix test module synchronization for network namespaces
To use the test modules in network namespaces, the test module
synchronization must be network namespace aware.

This patch extends the send_class() and sync_resource() methods with netns
argument to specify to which network namespace the module should be sent.

The patch also includes additional fix for the run_job() method. The
problem is that this method is called with an instance of a test module
as argument, this instance is pickled and sent to the slave where it's
unpickled. If the test module is sent to a network namespace the
ConnectionHandler in the initial namespace attempts to unpickle the message
which fails with ModuleNotFound error. This is because the test module
has been synchronized only for the network namespace. To fix this,
additional send_class call must be done for the initial namespace.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>