IpAddress.py: Change TODO to be to replace with ipaddress module
I happened upon this todo when looking at this code for MPTCP.

IMHO we should probably replace this with the builtin ipaddress module, or perhaps a customized subclass of it.

The ipaddress module contains many tools that can generate address, for example iterating though a subnet, and others.

https://docs.python.org/3/library/ipaddress.html#module-ipaddress

Refactoring this might be a lot of work, so I don't want to do it now, but I figured I would atleast update this todo as reminder to anyone who might be looking at the IpAddress code and want to do it.

Signed-off-by: Perry Gagne <pgagne@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>