Recipes.ENRT.BasePvPRecipe: fix VirtioType enum definition
The functional use of the Enum base class causes problems with
attempting to pickle the enum values. This is due to how the pickling
module tries to figure out which module the enum was defined in. This
can be improved by adding an explicit `module=` kwarg to the function
call; however, an additional issue is that the "VirtType" name needs to
also be identical to the "VirtioType" name that the object is being
assinged to.

As such there are two issues with the functional definiton of an enum.
I'm replacing this with the class based definition of an Enum as it
resolves both issues and is more easily understood and should be exactly
equivalent to the previous definition.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>