RecipeCommon: add L2TPManager
This adds L2TPManager class that can be used to create L2TP tunnels.
Due to the nature of the L2TP subsystem this can't be implemented as a
SoftDevice class. Instead user should use the init_class() method of the
Host API to create a machine specific instance of the class. Then the
instance methods create_tunnel(), delete_tunnel() can be used to manage
the L2TP tunnels.

Support for the L2TP session devices will be added in a followup patch.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>