Devices.Device: fix RTM_NEWADDR handling
The handler of RTM_NEWADDR incorrectly fetches the IFA_ADDRESS
attribute to update the device address. This is not an issue for IPv6 or
typical update of the device address. However it is a bug for the
assignment of address with peer address included (typical for point to
point networks), for example:

ip address add 192.168.100.1/24 peer 192.169.100.2 dev enp7s0

In this case, the IFA_ADDRESS attribute value is the peer address value
and the device address is the value of IFA_LOCAL.

When the peer address is not specified, both IFA_LOCAL and IFA_ADDRESS
contain the same value which is the reason this has been working without
any issues previously.

The fix is to use the IFA_LOCAL attribute value to update the device
address for IPv4 scenarios.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>