Controller.Machine: fix moving a device to a network namespace
When a device is moved to a network namespace an RTM_NEWLINK netlink
message is generated and handled by InterfaceManager. The InterfaceManager
creates a new instance of a generic Device class instead of a specific device
class (e.g. VethDevice). This device also remains in disabled state so
no actions or configurations can be made.

This patch fixes that by following changes.

The InterfaceManager is extended with remap_device() method that
creates a new instance of the specific device class that replaces the
generic Device instance.

The SlaveMethods are extended with remap_device() method that serves as
an rpc interface for calling InterfaceManager's method from the
Controller.

The Machine.remote_device_set_netns() method is updated to include an
explicit rpc_call to remap the device to proper type.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>