Devices: add LoopbackDevice
This patch adds implementation of a device class for the loopback device (lo)
so that users can benefit from the common API for the other devices, for
example IP address configuration.

The down() method is overridden to disallow the operation for this type of
device as this breaks connection between the controller and LNST agents
during recipe execution.

v2:

Removed the override of the up() method. This method is essential to use
the loopback device in a namespace because by default the device is down.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>