Controller.Machine: make device_database namespace aware
The Machine's device_database attribute contained devices from all namespaces
and potentially could be corrupted when two devices from different namespace
but with the same ifindex were added.

This patch modifies the device_database structure where the keys are individual
namespaces and values are dictionaries of the namespace devices.

Several modules that are accessing the device database are updated to be
compatible with the changes.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>