Controller.Namespace: update device's _id in device_database in special cases
This is to fix an issue with device id logged as None instead of a
handle defined by user. The problem is that the logger fetches the _id of the
corresponding RemoteDevice instance retrieved from the device_database instead
of the instance created by the user.

The RemoteDevice instance in the device_database is created independently
in following two cases:
 * a device is moved to a namespace - device_create event is called
 * a loopback device is created by user - here the LoopbackDevice already
   exists in the database after the InterfaceManager is initialized

In both cases, the device_database instances of the devices have _id set to
None and need to updated with the name defined by user.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>