Controller/Slave: handle RTM_DELLINK on net namespace changes
When a device is moved to a network namespace the original netns
receives RTM_DELLINK netlink message for the device. Handler of that
message will send dev_deleted event notification to Controller and
the Controller would set deleted flag for the device. Such device cannot
be manipulated anymore. This is fine for devices not being moved between
the namespaces. However when user moves a device this causes an issue.

For example:

    m1.gre = GreDevice(remote="192.168.122.2")
    m1.newns = NetNamespace("newns")
    m1.newns.gre = m1.gre
    m1.newns.gre.up()

The m1.gre is the device instance where the delete flag is set when the
assignment to m1.newns.gre is done. Since the m1.newns.gre and m1.gre are
the same instances any device manipulation, e.g. up/down() would result in
error.

To fix this, additional check is added to RTM_DELLINK event handler. When the
event is related to network namespace change then dev_netns_changed event
notification is sent to the Controller instead of dev_deleted event.

The dev_netns_changed notification triggers the same action on the Machine
instance as the dev_deleted one except setting the delete flag.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>