Devices.Device: handle RTM_NEWLINK messages without any IFLA_ADDRESS
There are cases when IFLA_ADDRESS is not included in the netlink
message, for example when vxlan device is created with gpe flag:

ip link add vxlan0 type vxlan external gpe

When the IFLA_ADDRESS is not present in the message the None value
is passed to hwaddress() parser function that raises an exception.

A check is added to hwaddr property whether the IFLA_ADDRESS attribute
is present. If it's not the parser function is bypassed and None is
returned.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>