Controller.Machine: set cacheonly mode for all devices on a machine cleanup
The RecipeRun object may be post evaluated, for example the results may be exported
into an external service. Accessing description property of some of the results may
access RemoteDevice attributes that would result in rpc calls to lnst agents
that may have already deleted the devices. To avoid this, all RemoteDevice
instances are set to be in readonly_cache mode.

The readonly_cache mode is also set on device_delete event.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>