Devices.VethDevice: peer property fixes
This patch fixes an issue when a VethDevice is moved to a namespace.

Currently, to get the peer property, VethDevice uses the instance of InterfaceManager
of the namespace the device was created in. This instance does not track
the peer device when it is moved to another namespace. When the peer property is
accessed, the InterfaceManager would raise a DeviceNotFound exception because the
peer device is no longer in the namespace.

To fix this issue, the peer property is renamed to peer_if_id property
that provides an information of the peer device ifindex accessible
through netlink API. The peer property will be implemented through
RemoteDevice in follow up patches.

Additionally the peer_name property is changed so that it can be used only when
the device is created through the netlink. Accessing the property value directly
is explicitly forbidden. Instead the value will be accessible through the peer_name
property of the RemoteDevice instance from a recipe.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>