Controller.Machine: handle device namespace movement
When a user moves a device to a namespace, the remote_device_set_netns method
moves the device from device_database to the temporary cache of devices.

When the action is performed on the LNST agent, the notification handlers
device_netns_change (RTM_DELLINK) and device_created (RTM_NEWLINK) uses this
cache to lookup and update the device.

Signed-off-by: Jan Tluka <jtluka@redhat.com>
Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>