lnst.Tests.TestPMD: adapt to two executable names
Added a property that returns the name of the base testpmd executable.

At some point in the dpdk upstream development the name of the binary
changed to "dpdk-testpmd", breaking the use of this TestModule on rhel8.

This also adds an additional "sanity check" by testing to see if the
executable is actually installed with "is_installed" and raising an
exception in case it isn't.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>