Parameters.py: Add get for params object.
This adds a method to the Parameters object to let you more easily query for parameters that may or may not be in self.params objects.

It allows your to check for existence of a param, and return the value (or a default if it doesn't exist). Its similar to dict.get.