FlowMeasurementGenerator.py: New combined MeasurementGenerator for flow traffic.
This patch creates a unified FlowMeasurementGenerator that replaces the tool specific IperfMeasurementGenerator and NeperMeasurementGenerator. Most of the code was duplicated in both, and the minor differences where able to be handled by the using the `net_perf_tool` parameter.

The default is to use iperf. If a recipe wishes to specify which tool it uses, it should override the `net_perf_tool` parameter and set it to the name of the tool (see MEASUREMENT_LOOKUP)

The new FlowMeasurementGenerator was mostly based on IperfMeasurementGenerator.

NeperFlowMeasurement also needs some changes (mostly copied from IperfFlowMeasurenet) to be compatable with FlowMeasurementGenerator. It the future we might want to look to also unify NeperFlowMeasurement and IperfFlowMeasurement.

This unified Measurement generator can also be used further used for future tools (netperf for example) if need be.