ShortLivedConnectionsRecipe.py: Update to use FlowMeasurementGenerator.
ShortLivedConnectionsRecipe can now be based on the unified FlowMeasurementGenerator, instead of a Neper specific one.

The recipe overrides the net_perf_tool parameter to only allow for for `neper` to be used.

Due to this unification there wasn't much of a need to keep LatencyEnrtRecipe, so ShortLivedConnectionsRecipe is now just subclassed directly from BaremetalEnrtRecipe.