SimpleMacsecRecipe: remove generate_perf_configurations method
After a bunch of refactoring commits in history this should now be a
redundant method. At the same time the recent refactoring of the
FlowMeasurementGenerator method this code is broken as the semantics of
the `self.params.net_perf_tool` parameter have changed.

Removing this method should resolve the issue and simplify the code.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>