lnst: add pkgutil style namespace packages
Removing top level init that was redundant and not used.

Updating the pyproject.toml, bumping project
version to 16 and python version to 3.6 and higher
to speed up installation.

Whenever LNST was installed as a dependent project
poetry was unable to stitch together multiple
packages and their modules.
Easiest solution is to update paths accordingly [0].

[0] - https://packaging.python.org/guides/packaging-namespace-packages/#pkgutil-style-namespace-packages