ENRT.DevInterruptHWConfigMixin: refactor irq pinning
After some discussion with Paolo Abeni, the /proc/interrupt file should
be the source of information for the device irqs that we need to be
pinning to specific cpus. This is because it contains the "in use" list
of irqs in contrast with the "msi_irqs" file which contains the
"reserved" but potentially "not in use" irqs.

At the same time for the /proc/interrupt method to work properly the
device needs to be up when we read the file. I've added this
functionality to the irq number list detection method.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>