Tests/Iperf: fix parsing of empty json output
Iperf running as server may exit with return code 0 and return an empty json.
This usually happens when the iperf running as client cannot complete
the test, for example when running a UDP stream test with multiple
streams on debug kernel.

If the server's json output cannot be parsed the test module should
return False status.

Signed-off-by: Jan Tluka <jtluka@redhat.com>