Fix copy_file_from_machine inifinite loop
Controller's copy_file_from_machine was comparing received bytes with an
empty python string. But empty bytes and empty string are not equal in
python, so the comparison failed and resulted in an endless while loop.

Also changed copy_file_to_machine's check in order to unify both of
them.